AI Capital Estimator -v2/backend/requirements.txt
fastapi==0.134.0
uvicorn[standard]==0.23.2
pydantic==2.9.0
python-dotenv==1.0.0
openai==1.3.0
